requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
djangorestframework==3.3.0
plaid-python==1.1.0
djoser==0.4.0
progressbar2==3.5.0
